moved new controller logic
